requirements.txt
Django>=4.2,<5.1
djangorestframework==3.16.1
gunicorn>=21.2
openai>=1.0
requests>=2.30
django-cors-headers>=4.0
python-dotenv>=1.0
whitenoise>=6.6
google-generativeai>=0.5.0